refactor: change Api classes
