+++ src/api/index.ts
@@ -1,0 +1,5 @@
+import Users from './classes/Users';
+import Auth from './classes/Auth';
+
+export const users = new Users();
+export const auth = new Auth();
